Converting `generators` into `constructors`
