+++ index.js
@@ -8,12 +8,6 @@ const fmt = util.format.bind(util)
 const dbProto = {}
 const tableProto = {}
 
-var defaultGenerator = function generateTableRow(data) {
-  // somewhat naive - potentially worth checking data for correctness
-  // currently assuming that data passed in is correct for table
-  return create(this.row, data)
-}
-
 dbProto.close = function close(callback) {
   return this.driver.close(this.connection, callback)
 }
@@ -35,17 +29,22 @@ dbProto.registerTable = function registerTable(name, def) {
   if (fields.indexOf(primaryKey) === -1)
     fields.unshift(primaryKey)
 
-  var generator = def.generator || defaultGenerator
+  if (!def.hasOwnProperty('constructor')) {
+    def.constructor = function createRow (data) {
+      return create(table.row, data)
+    }
+  }
 
   const table = create(tableProto, {
     table: def.tableName || name,
     primaryKey: primaryKey,
     fields: fields,
     row: def.methods || {},
-    create: generator,
+    constructor: def.constructor,
     relationships: def.relationships || {},
     db: this,
   })
+
   this.tables[name] = table
   return table
 }
@@ -102,7 +101,7 @@ tableProto.get = function get(cnd, opts, callback) {
     opts = {}
   }
 
-  const createRow = this.create.bind(this)
+  const RowClass = this.constructor
   const driver = this.db.driver
 
   const table = this.table
@@ -113,7 +112,6 @@ tableProto.get = function get(cnd, opts, callback) {
       relationships === true) {
     relationships = this.relationships
   }
-
 
   const selectSql = driver.selectSql({
     db: this.db,
@@ -150,7 +148,7 @@ tableProto.get = function get(cnd, opts, callback) {
     driver.hydrateRow(singleton, hydrOpts, function (err, result) {
       if (err) return callback(err)
 
-      return callback(null, createRow(result))
+      return callback(null, new RowClass(result))
     })
   }
 
@@ -161,7 +159,7 @@ tableProto.get = function get(cnd, opts, callback) {
     driver.hydrateRows(rows, hydrOpts, function (err, rows) {
       if (err) return callback(err)
       return callback(null, rows.map(function (row) {
-        return createRow(row)
+        return new RowClass(row)
       }))
     })
   }
