+++ lib/drivers/prototype.js
@@ -218,7 +218,7 @@ module.exports = {
     forEach(relationships, function (rel, relKey) {
       if (rel.type !== 'hasOne') return
       const foreign = tableCache[rel.foreign.table]
-      row[relKey] = foreign.create(row[relKey])
+      row[relKey] = new foreign.constructor(row[relKey])
     })
 
     return row
@@ -259,7 +259,7 @@ module.exports = {
         }
 
         globalRow[relKey] = rows.map(function (row) {
-          return foreign.create(row)
+          return new foreign.constructor(row)
         })
 
         return checkDone()
